requirements.txt
#--------------------------------------
# Required packages to install via pip
#--------------------------------------

# Flask framework
Flask

# Turso API library
libsql-client

# Environment Variables
python-dotenv

# Coloured text
colorama

# Timezones
tzdata
